Made the Signup Screen
This Screen consists of the First Name field, The last name field, The user name, The password and an option below that allows the user to login instead if the already have a password. There is an issue with the login function that has to be addressed. The link to the login page does not take you to the login screen due to a problem with the navigation code.